Refactor to use shared code between layer types
